refactor: extract scenario governance policy
