+++ src/app/services/scenario_engine.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import datetime as dt
 import hashlib
 import json
 from dataclasses import dataclass
@@ -10,13 +9,15 @@ from app.contracts.scenario import (
     RegimeScenarioPackResponse,
     ScenarioEvaluationMetadata,
     ScenarioExposureComponent,
-    ScenarioPackApplicabilityStatus,
-    ScenarioPackApprovalStatus,
-    ScenarioPackEffectivePeriodStatus,
     ScenarioPackGovernanceEvidence,
     ScenarioPositionContribution,
     ScenarioResult,
     ScenarioSupportabilityState,
+)
+from app.services.scenario_governance import (
+    SCENARIO_PACK_GOVERNANCE as SCENARIO_PACK_GOVERNANCE,  # noqa: F401
+    evaluate_governance_evidence,
+    most_severe_supportability,
 )
 
 
@@ -28,46 +29,11 @@ class ScenarioDefinition:
 
 
 @dataclass(frozen=True)
-class ScenarioPackGovernanceDefinition:
-    cio_approval_status: ScenarioPackApprovalStatus
-    cio_approval_ref: str
-    approved_by: str
-    approved_at: dt.datetime
-    effective_from: dt.date
-    effective_to: dt.date
-    applicability_scope: tuple[str, ...]
-    applicable_portfolio_ids: frozenset[str]
-    methodology_ref: str
-
-
-@dataclass(frozen=True)
 class _ScenarioPackEvaluation:
     scenario_results: list[ScenarioResult]
     worst_case_loss: float
     breach: bool
     governance_evidence: ScenarioPackGovernanceEvidence
-    reason_codes: list[str]
-    supportability: ScenarioSupportabilityState
-
-
-@dataclass(frozen=True)
-class _EffectivePeriodDecision:
-    status: ScenarioPackEffectivePeriodStatus
-    reason_codes: list[str]
-    supportability: ScenarioSupportabilityState
-
-
-@dataclass(frozen=True)
-class _PortfolioApplicabilityDecision:
-    status: ScenarioPackApplicabilityStatus
-    portfolio_applicability_ref: str | None
-    reason_codes: list[str]
-    supportability: ScenarioSupportabilityState
-
-
-@dataclass(frozen=True)
-class _GovernanceEvaluation:
-    evidence: ScenarioPackGovernanceEvidence
     reason_codes: list[str]
     supportability: ScenarioSupportabilityState
 
@@ -107,19 +73,6 @@ SCENARIO_PACKS: dict[str, tuple[ScenarioDefinition, ...]] = {
     )
 }
 SUPPORTED_BUCKETS = frozenset({"EQUITY", "FIXED_INCOME", "ALTERNATIVES", "CASH"})
-SCENARIO_PACK_GOVERNANCE: dict[str, ScenarioPackGovernanceDefinition] = {
-    "CIO_REGIME_2026_Q2": ScenarioPackGovernanceDefinition(
-        cio_approval_status=ScenarioPackApprovalStatus.APPROVED,
-        cio_approval_ref="CIO-REGIME-2026-Q2-APPROVAL",
-        approved_by="CIO Risk Committee",
-        approved_at=dt.datetime(2026, 4, 15, 9, 0, tzinfo=dt.UTC),
-        effective_from=dt.date(2026, 4, 1),
-        effective_to=dt.date(2026, 6, 30),
-        applicability_scope=("DISCRETIONARY_PRIVATE_BANKING_BALANCED",),
-        applicable_portfolio_ids=frozenset({"PB_SG_GLOBAL_BAL_001"}),
-        methodology_ref="docs/methodologies/metrics/regime-scenario-pack-evaluation.md",
-    )
-}
 
 
 def _evaluate_scenario_pack_context(
@@ -135,8 +88,8 @@ def _evaluate_scenario_pack_context(
         if unsupported_buckets
         else ScenarioSupportabilityState.READY
     )
-    governance = _evaluate_governance_evidence(request)
-    supportability = _most_severe_supportability(
+    governance = evaluate_governance_evidence(request)
+    supportability = most_severe_supportability(
         supportability,
         governance.supportability,
     )
@@ -194,120 +147,6 @@ def evaluate_regime_scenario_pack(
             calculation_supportability=evaluation.supportability,
         ),
     )
-
-
-def _effective_period_decision(
-    *,
-    governance: ScenarioPackGovernanceDefinition,
-    as_of_date: dt.date,
-) -> _EffectivePeriodDecision:
-    if as_of_date < governance.effective_from:
-        return _EffectivePeriodDecision(
-            status=ScenarioPackEffectivePeriodStatus.NOT_YET_EFFECTIVE,
-            reason_codes=["REGIME_SCENARIO_EFFECTIVE_PERIOD_EXCEPTION"],
-            supportability=ScenarioSupportabilityState.DEGRADED,
-        )
-    if as_of_date > governance.effective_to:
-        return _EffectivePeriodDecision(
-            status=ScenarioPackEffectivePeriodStatus.EXPIRED,
-            reason_codes=["REGIME_SCENARIO_EFFECTIVE_PERIOD_EXCEPTION"],
-            supportability=ScenarioSupportabilityState.DEGRADED,
-        )
-    return _EffectivePeriodDecision(
-        status=ScenarioPackEffectivePeriodStatus.ACTIVE,
-        reason_codes=[],
-        supportability=ScenarioSupportabilityState.READY,
-    )
-
-
-def _portfolio_applicability_decision(
-    *,
-    governance: ScenarioPackGovernanceDefinition,
-    portfolio_id: str | None,
-) -> _PortfolioApplicabilityDecision:
-    if portfolio_id is None:
-        return _PortfolioApplicabilityDecision(
-            status=ScenarioPackApplicabilityStatus.PENDING_REVIEW,
-            portfolio_applicability_ref=None,
-            reason_codes=["REGIME_SCENARIO_PORTFOLIO_APPLICABILITY_NOT_CONFIRMED"],
-            supportability=ScenarioSupportabilityState.PENDING_REVIEW,
-        )
-    if portfolio_id in governance.applicable_portfolio_ids:
-        return _PortfolioApplicabilityDecision(
-            status=ScenarioPackApplicabilityStatus.APPLICABLE,
-            portfolio_applicability_ref=f"{governance.cio_approval_ref}-APP-{portfolio_id}",
-            reason_codes=[],
-            supportability=ScenarioSupportabilityState.READY,
-        )
-    return _PortfolioApplicabilityDecision(
-        status=ScenarioPackApplicabilityStatus.NOT_APPLICABLE,
-        portfolio_applicability_ref=None,
-        reason_codes=["REGIME_SCENARIO_PORTFOLIO_NOT_APPLICABLE"],
-        supportability=ScenarioSupportabilityState.BLOCKED,
-    )
-
-
-def _evaluate_governance_evidence(
-    request: RegimeScenarioPackRequest,
-) -> _GovernanceEvaluation:
-    governance = SCENARIO_PACK_GOVERNANCE[request.scenario_pack_id]
-    reason_codes: list[str] = []
-    supportability = ScenarioSupportabilityState.READY
-
-    if governance.cio_approval_status != ScenarioPackApprovalStatus.APPROVED:
-        reason_codes.append("REGIME_SCENARIO_CIO_APPROVAL_NOT_CONFIRMED")
-        supportability = ScenarioSupportabilityState.BLOCKED
-
-    effective_period = _effective_period_decision(
-        governance=governance,
-        as_of_date=request.as_of_date,
-    )
-    reason_codes.extend(effective_period.reason_codes)
-    supportability = _most_severe_supportability(
-        supportability,
-        effective_period.supportability,
-    )
-
-    portfolio_applicability = _portfolio_applicability_decision(
-        governance=governance,
-        portfolio_id=request.portfolio_id,
-    )
-    reason_codes.extend(portfolio_applicability.reason_codes)
-    supportability = _most_severe_supportability(
-        supportability,
-        portfolio_applicability.supportability,
-    )
-
-    return _GovernanceEvaluation(
-        evidence=ScenarioPackGovernanceEvidence(
-            cio_approval_status=governance.cio_approval_status,
-            cio_approval_ref=governance.cio_approval_ref,
-            approved_by=governance.approved_by,
-            approved_at=governance.approved_at,
-            effective_from=governance.effective_from,
-            effective_to=governance.effective_to,
-            effective_period_status=effective_period.status,
-            applicability_status=portfolio_applicability.status,
-            applicability_scope=list(governance.applicability_scope),
-            portfolio_applicability_ref=portfolio_applicability.portfolio_applicability_ref,
-            methodology_ref=governance.methodology_ref,
-        ),
-        reason_codes=reason_codes,
-        supportability=supportability,
-    )
-
-
-def _most_severe_supportability(
-    left: ScenarioSupportabilityState,
-    right: ScenarioSupportabilityState,
-) -> ScenarioSupportabilityState:
-    severity = {
-        ScenarioSupportabilityState.READY: 0,
-        ScenarioSupportabilityState.PENDING_REVIEW: 1,
-        ScenarioSupportabilityState.DEGRADED: 2,
-        ScenarioSupportabilityState.BLOCKED: 3,
-    }
-    return left if severity[left] >= severity[right] else right
 
 
 def _evaluate_scenario(
